create responsive table component
