+++ src/client/components/Thema/ThemaPaginaTable.tsx
@@ -7,12 +7,10 @@ import type {
   ZaakAanvraagDetail,
 } from '../../../universal/types/App.types.ts';
 import { MAX_TABLE_ROWS_ON_THEMA_PAGINA } from '../../config/app.ts';
-import { useSmallScreen } from '../../hooks/media.hook.ts';
 import { LinkToListPage } from '../LinkToListPage/LinkToListPage.tsx';
 import { PageContentCell } from '../Page/Page.tsx';
-import { TableMobileView } from '../Table/TableMobileView.tsx';
+import { ResponsiveTable } from '../Table/ResponsiveTable.tsx';
 import type { DisplayProps } from '../Table/TableV2.tsx';
-import { TableV2 } from '../Table/TableV2.tsx';
 
 const DISPLAY_PROPS_DEFAULT: DisplayProps<{ title: string }> = {
   title: 'Titel',
@@ -53,37 +51,19 @@ export function ThemaPaginaTable<
     ? `U heeft (nog) geen ${title.toLowerCase()}`
     : TEXT_NO_CONTENT_DEFAULT;
 
-  const isSmallScreen = useSmallScreen();
-  const isMobileListEnabledInTheme =
-    'enableMobileListView' in displayProps
-      ? displayProps.enableMobileListView
-      : false;
-
   const hasListPage = !!listPageRoute && maxItems !== -1;
 
   const zaken_ = hasListPage ? zaken.slice(0, maxItems) : zaken;
 
   return (
     <PageContentCell>
-      {isSmallScreen && isMobileListEnabledInTheme ? (
-        <TableMobileView<T>
-          displayProps={displayProps}
-          items={zaken_}
-          caption={title}
-          contentAfterTheCaption={contentAfterTheTitle}
-          className={className}
-        />
-      ) : (
-        <TableV2
-          showTHead={!!zaken.length}
-          caption={title}
-          contentAfterTheCaption={contentAfterTheTitle}
-          items={zaken_}
-          displayProps={displayProps}
-          className={className}
-        />
-      )}
-
+      <ResponsiveTable
+        items={zaken_}
+        caption={title}
+        contentAfterTheCaption={contentAfterTheTitle}
+        displayProps={displayProps}
+        className={className}
+      />
       {!zaken.length && (
         <Paragraph>{textNoContent ?? textNoContentDefault}</Paragraph>
       )}
